requirements.txt
pillow==11.0.0
rembg==2.0.28